+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.10.1
+
 ## 3.0.0
 
 ### Major Changes