+++ packages/domain-social/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.10.2
+
 ## 0.3.5
 
 ### Patch Changes