+++ packages/health-check-nest/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.28
+
+### Patch Changes
+
+- Updated dependencies
+- @volontariapp/errors-nest@0.10.1
+
 ## 0.1.27
 
 ### Patch Changes